Reorganise some files
