+++ src/daemon_communication/daemon_client.rs
