+++ src/daemon_communication/mod.rs
@@ -1,3 +1,4 @@
-pub mod client;
+pub mod app;
+pub mod daemon_client;
 pub mod server;
 pub mod structs;
